feat: move (BlueBack's) common auto functions to separate classes
