+++ IFtcOpmodeSleep.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode;
+
+
+public interface IFtcOpmodeSleep {
+  void FtcSleepFnc(int ms);
+}
